Resolveproperties 090421@1104 (#11)
* Add initial impl and tests for resolving properties.

* Leave resolve checks out of resolving methods against unwrapped types. See the second test in StringSpec.

* Merge two filters into one.

* Remove unused variable

* Test abstracts and classes against new ResolveProperty.hx resolver.

* Test generic parameters and start thinking how to fix/implement them.

* An improvement to detecting type parameters and using them to resolve matches.

* Add a check in catchAll of ResolveProperty.hx, as already processed types can be passed thru again. Not entirely sure why this started happening after improve type parameter handling.
Disable type parameter resolving in findProperty/filterFields as types should be resolve/sealed in determineTask, so all down stream functions **should** be working with known concrete types.

* Remove unneeded code.

* All tests now pass again.
I expect the main problem was caused by the old filter function not touching some* types with `unify`, which is used in the new filter function. Which worked! but causes new issues to arise.

* Remove old code.

* Merge `resolve` and `catchAll` methods into one.